Move Java tests into the new harness
