+++ packages/pigeon/platform_tests/alternate_language_test_plugin/android/build.gradle
@@ -32,4 +32,21 @@ android {
     defaultConfig {
         minSdkVersion 16
     }
+
+    testOptions {
+        unitTests.includeAndroidResources = true
+        unitTests.returnDefaultValues = true
+        unitTests.all {
+            testLogging {
+               events "passed", "skipped", "failed", "standardOut", "standardError"
+               outputs.upToDateWhen {false}
+               showStandardStreams = true
+            }
+        }
+    }
+
+    dependencies {
+        testImplementation 'junit:junit:4.+'
+        testImplementation "org.mockito:mockito-core:3.+"
+    }
 }
