+++ packages/pigeon/platform_tests/alternate_language_test_plugin/android/src/main/.gitignore
@@ -1,0 +1,5 @@
+# TODO(stuartmorgan): Remove this, so that review will show the effects of
+# changes on generated files. This will need a way to avoid unnecessary churn,
+# such as a flag to suppress version stamp generation.
+*.java
+!AlternateLanguageTestPlugin.kt
