--- packages/pigeon/platform_tests/android_unit_tests/.metadata
@@ -1,10 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
-
-version:
-  revision: 891511d58f6550ce9e9b03b8d7c6a602caa97488
-  channel: master
-
-project_type: app
