--- packages/pigeon/platform_tests/android_unit_tests/README.md
@@ -1,3 +1,0 @@
-# android_unit_tests
-
-Unit-tests for Pigeon generated Java code.  See [../../run_tests.sh](../../run_tests.sh).
