--- packages/pigeon/platform_tests/android_unit_tests/android/app/src/main/java/com/example/android_unit_tests/.gitignore
@@ -1,1 +1,0 @@
-*.java
