--- packages/pigeon/platform_tests/android_unit_tests/lib/main.dart
@@ -1,5 +1,0 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-void main() {}
